#1220 review: move Task.ts to `types/task`
