+++ wls-gui-wahllokalsystem/src/components/common/progressLinear/BaseProgressLinear.vue
@@ -24,7 +24,7 @@
   </div>
 </template>
 <script setup lang="ts">
-import type { Task } from "@/types/Task.ts";
+import type { Task } from "@/types/tasks/Task.ts";
 
 import {
   VExpansionPanel,
