+++ wls-gui-wahllokalsystem/src/types/tasks/Task.ts
